faiss_builder/requirements.txt
duckdb
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pyyaml
fastapi
uvicorn